114 - layouts, pages, components, routerの構造
問題
現在の構造だと何回も繰り返しているコードがある。

解決策
ルーターのネスティングを使って、重複しているコードを削減。ネスティングをに切り替えて後の問題（Sliderは正しく動かない）を解決。
