--- imports/ui/layouts/AddSmartRemote.jsx
@@ -1,90 +1,0 @@
-import React from 'react';
-import i18n from 'meteor/universe:i18n';
-import Header from '../components/Header';
-import Validation from 'react-validation';
-import '../../api/validator/form_validator';
-import TrackerReact from 'meteor/ultimatejs:tracker-react';
-import {Grid, Col, Row } from 'react-bootstrap';
-import Alert from 'react-s-alert';
-import 'react-s-alert/dist/s-alert-default.css';
-import 'react-s-alert/dist/s-alert-css-effects/genie.css';
-import { browserHistory } from 'react-router';
-import Slider from '../components/Slider';
-
-export default class AddSmartRemote extends TrackerReact(React.Component) {
-  constructor(props) {
-    super(props);
-    this.state = {
-      isAuthenticated: Meteor.userId() !== null,
-    };
-  }
-
-  componentWillMount() {
-    if (!this.state.isAuthenticated) {
-      browserHistory.push('/sign-in');
-    }
-  }
-
-  handleSubmit(event) {
-    event.preventDefault();
-    const params = {
-      deviceId: event.target.deviceId.value,
-      accessToken: event.target.accessToken.value,
-    };
-    Meteor.call('addPhotonInfo', params, (error)=>{
-      if (!error) {
-        document.getElementById('input-device-id').value = '';
-        document.getElementById('input-access-token').value = '';
-        Alert.success(i18n.getTranslation('addPhoton', 'success'), {
-          position: 'bottom',
-          effect: 'genie',
-          timeout: 3000,
-        });
-      }
-    });
-  }
-
-  render() {
-    if (Meteor.user()) {
-      return (
-        <div>
-          <Header/>
-          <Slider/>
-          <Grid className="center button-register-center">
-            <Validation.components.Form onSubmit={this.handleSubmit.bind(this)} >
-              <Row>
-                <Col>
-                  <Validation.components.Input
-                    id='input-device-id'
-                    className='input-style'
-                    type='text'
-                    name='deviceId'
-                    value=''
-                    placeholder={i18n.getTranslation('addPhoton', 'deviceId')}
-                    validations={['required', 'deviceId']}/>
-                  <Validation.components.Input
-                    id='input-access-token'
-                    className='input-style'
-                    type='text'
-                    name='accessToken'
-                    value=''
-                    placeholder={i18n.getTranslation('addPhoton', 'accessToken')}
-                    validations={['required', 'accessToken']}/>
-                </Col>
-              </Row>
-              <Row>
-                <Col>
-                  <Validation.components.Button className='button-style add-photon-button'>{i18n.getTranslation('addPhoton', 'register')}</Validation.components.Button>
-                </Col>
-              </Row>
-            </Validation.components.Form>
-          </Grid>
-          <Alert stack={{limit: 1}} />
-        </div>
-      );
-    } else {
-      return null;
-    }
-
-  }
-}
