+++ imports/ui/layouts/LoggedOut.jsx
@@ -1,10 +1,13 @@
 import React from 'react';
 import i18n from 'meteor/universe:i18n';
 import Header from '../../ui/components/Header';
-import { Link } from 'react-router';
 import { browserHistory } from 'react-router';
 import TrackerReact from 'meteor/ultimatejs:tracker-react';
-export default class Top extends TrackerReact(React.Component) {
+import Alert from 'react-s-alert';
+import 'react-s-alert/dist/s-alert-default.css';
+import 'react-s-alert/dist/s-alert-css-effects/genie.css';
+
+export default class LoggedOut extends TrackerReact(React.Component) {
   constructor() {
     super(...arguments);
     this.state = {
@@ -30,17 +33,12 @@ export default class Top extends TrackerReact(React.Component) {
   }
 
   render() {
+    const childrenWithProps = React.cloneElement(this.props.children, this.props);
     return (
       <div>
         <Header/>
-        <div className="top-box">
-          <h1>{i18n.getTranslation('top', 'title1')}</h1>
-          <h1>{i18n.getTranslation('top', 'title2')}</h1>
-          <h3>{i18n.getTranslation('top', 'subTitle')}</h3>
-        </div>
-        <div className="top-button">
-          <Link to="sign-in" className="sign-in button-style top-button-padding">{i18n.getTranslation('top', 'signIn')}</Link>
-        </div>
+        {childrenWithProps}
+        <Alert stack={{limit: 1}} />
       </div>
     );
   }
